pydrake: Rename pybind utility headers.
