+++ bindings/pydrake/BUILD.bazel
@@ -25,6 +25,14 @@ package(default_visibility = adjust_labels_for_drake_hoist([
 
 # N.B. Any common headers should be shared via a HEADER ONLY library.
 
+# Target Naming Conventions:
+# `*_py`
+#   A Python library (can be pure Python or pybind)
+#   Files: `*.py`, `*_py.cc`
+# `*_pybind`
+#   A C++ library for adding pybind-specific utilities to be consumed by C++.
+#   Files: `*_pybind.{h,cc}`
+
 # This is a pure-python, standalone library.
 # Keep this away from `common_py` to simplify test dependencies.
 drake_py_library(
@@ -47,8 +55,8 @@ drake_pybind_library(
 
 # TODO(eric.cousineau): Ensure that this is not installed.
 drake_cc_library(
-    name = "autodiff_types",
-    hdrs = ["autodiff_types_py.h"],
+    name = "autodiff_types_pybind",
+    hdrs = ["autodiff_types_pybind.h"],
     # NOTE: We must violate IWYU and not list "autodiff" in "deps" here because
     # we do not want these dependencies leaking in at install time.
 )
@@ -56,7 +64,7 @@ drake_cc_library(
 drake_pybind_library(
     name = "autodiffutils_py",
     cc_deps = [
-        ":autodiff_types",
+        ":autodiff_types_pybind",
     ],
     cc_srcs = ["autodiffutils_py.cc"],
     py_deps = [
@@ -71,7 +79,7 @@ drake_pybind_library(
 drake_pybind_library(
     name = "rbtree_py",
     cc_deps = [
-        ":autodiff_types",
+        ":autodiff_types_pybind",
     ],
     cc_srcs = ["rbtree_py.cc"],
     py_deps = [
@@ -93,14 +101,14 @@ drake_pybind_library(
 
 # TODO(eric.cousineau): Make private.
 drake_cc_library(
-    name = "symbolic_types_py_h",
-    hdrs = ["symbolic_types_py.h"],
+    name = "symbolic_types_pybind",
+    hdrs = ["symbolic_types_pybind.h"],
 )
 
 drake_pybind_library(
     name = "symbolic_py",
     cc_deps = [
-        ":symbolic_types_py_h",
+        ":symbolic_types_pybind",
     ],
     cc_srcs = ["symbolic_py.cc"],
     py_deps = [
@@ -143,7 +151,7 @@ drake_pybind_library(
     testonly = 1,
     add_install = False,
     cc_deps = [
-        ":symbolic_types_py_h",
+        ":symbolic_types_pybind",
     ],
     cc_so_name = "test/odr_test_module",
     cc_srcs = ["test/odr_test_module_py.cc"],
