+++ bindings/pydrake/symbolic_py.cc
@@ -2,7 +2,7 @@
 #include <pybind11/operators.h>
 #include <pybind11/pybind11.h>
 
-#include "drake/bindings/pydrake/symbolic_types_py.h"
+#include "drake/bindings/pydrake/symbolic_types_pybind.h"
 
 namespace py = pybind11;
