+++ bindings/pydrake/symbolic_types_pybind.h
